requirements.txt
Flask==3.1.0
tensorflow==2.19.0
keras==3.5.0
numpy==1.26.0
python-dotenv==1.1.0
wikipedia-api==0.8.1
google-generativeai==0.8.4
mysql-connector-python==9.2.0
gunicorn==23.0.0